refactor(cli): split vindex3_cmd/lowered.rs into lowered/{mod,resident,routed,dump} under the 800-line cap
No behaviour change: resident-weight loading, the routed (expert-bank)
FFN builder and the --dump-layers writer move into their own files;
mod.rs keeps the session and run_lowered.
