+++ crates/larql-cli/src/commands/primary/vindex3_cmd/lowered/mod.rs
@@ -31,27 +31,33 @@ use larql_vindex::error::VindexError;
 use larql_vindex::format::vindex3::graph::policy::AttentionSpan;
 use larql_vindex::format::vindex3::opplan::exec::backend::{WeightFormat, WeightFormats};
 use larql_vindex::format::vindex3::opplan::exec::operands::OperandStore;
-use larql_vindex::format::vindex3::opplan::exec::weights::{
-    load_weight, AlignedBytes, LoadedWeight,
-};
-use larql_vindex::format::vindex3::opplan::{
-    ComponentOpPlan, FfnOp, LayerPlan, NormOp, OperandRef,
-};
+use larql_vindex::format::vindex3::opplan::exec::weights::LoadedWeight;
+use larql_vindex::format::vindex3::opplan::{ComponentOpPlan, LayerPlan};
 
 /// One matrix operand, resident on the device.
-struct DeviceMatrix {
+mod dump;
+mod resident;
+mod routed;
+
+use dump::dump_lowered;
+use resident::{
+    resident_matrix, resident_norm, resident_vector, rope_inv_freq_table, rope_table_key, Ablation,
+};
+use routed::{build_ffn, FfnResident};
+
+pub(super) struct DeviceMatrix {
     /// `scales` is unused for f16; the representation is what the plan's
     /// per-class policy asked for, not something inferred here.
-    packed: DeviceBuffer,
-    scales: DeviceBuffer,
-    tensor_scale: f32,
-    format: WeightFormat,
-    rows: usize,
-    cols: usize,
+    pub(super) packed: DeviceBuffer,
+    pub(super) scales: DeviceBuffer,
+    pub(super) tensor_scale: f32,
+    pub(super) format: WeightFormat,
+    pub(super) rows: usize,
+    pub(super) cols: usize,
 }
 
 impl DeviceMatrix {
-    fn as_lowered(&self) -> LoweredMatrix<'_> {
+    pub(super) fn as_lowered(&self) -> LoweredMatrix<'_> {
         match self.format {
             WeightFormat::F16 => LoweredMatrix::F16 {
                 bytes: &self.packed,
@@ -91,135 +97,6 @@ struct LayerResident {
     v_cache: DeviceBuffer,
 }
 
-/// A layer's resident FFN: dense gate/up/down matrices, or a routed
-/// expert bank resolved to registered regions plus the served MoE
-/// scratch and descriptor table.
-enum FfnResident {
-    Dense {
-        gate: DeviceMatrix,
-        up: DeviceMatrix,
-        down: DeviceMatrix,
-    },
-    Routed(Box<RoutedLayer>),
-}
-
-/// One routed layer's expert bank and MoE machinery, held for the
-/// session: the packed expert bytes in page-aligned, region-registered
-/// buffers (bound zero-copy, never copied per token), the f32 router and
-/// bias operands, the per-layer `MoeScratch`, and the descriptor table.
-struct RoutedLayer {
-    gate_up_blocks: AlignedBytes,
-    gate_up_scales: AlignedBytes,
-    down_blocks: AlignedBytes,
-    down_scales: AlignedBytes,
-    router_proj: Vec<f32>,
-    router_bias: Vec<f32>,
-    gate_up_bias: Vec<f32>,
-    down_bias: Vec<f32>,
-    pre_ffn_norm: Vec<f32>,
-    gu_expert_bytes: usize,
-    gu_scale_bytes: usize,
-    dn_expert_bytes: usize,
-    dn_scale_bytes: usize,
-    experts: usize,
-    top_k: usize,
-    inter: usize,
-    gate_rule: larql_compute::MoeGateRule,
-    table: std::sync::Arc<larql_compute_metal::moe_descriptor::MoeExpertDescriptorTable>,
-    scratch: larql_compute_metal::MoeScratch,
-    eps: f32,
-}
-
-/// Per-expert byte slices into a packed bank: expert `e` occupies
-/// `[e*per .. (e+1)*per]` of the bank's logical bytes.
-fn expert_slices(bank: &AlignedBytes, per: usize, experts: usize) -> Vec<&[u8]> {
-    let all = &bank.as_slice()[..per * experts];
-    (0..experts).map(|e| &all[e * per..(e + 1) * per]).collect()
-}
-
-impl RoutedLayer {
-    /// The `MoeLayerWeights` view the served descriptor path consumes,
-    /// assembled from a `RoutedFfnOp` — per-expert slices into the
-    /// registered banks, router/bias from f32 storage, GPT-OSS routing
-    /// and gate semantics from the plan. Rebuilt per step (borrows are
-    /// cheap; no bytes move).
-    fn moe(&self) -> larql_compute::MoeLayerWeights<'_> {
-        use larql_compute::{
-            MoeExpertScales, MoeFusedRowLayout, MoeRoutingPolicy, MoeWeightLayout, QuantFormat,
-        };
-        larql_compute::MoeLayerWeights {
-            experts_gate_up: expert_slices(
-                &self.gate_up_blocks,
-                self.gu_expert_bytes,
-                self.experts,
-            ),
-            experts_down: expert_slices(&self.down_blocks, self.dn_expert_bytes, self.experts),
-            routing_policy: MoeRoutingPolicy::top_k_then_softmax(),
-            weight_layout: MoeWeightLayout::unpadded(),
-            expert_scales: MoeExpertScales::Paired {
-                gate_up: expert_slices(&self.gate_up_scales, self.gu_scale_bytes, self.experts),
-                down: expert_slices(&self.down_scales, self.dn_scale_bytes, self.experts),
-            },
-            fused_row_layout: MoeFusedRowLayout::Interleaved,
-            expert_data_format: QuantFormat::MXFP4,
-            router_proj: &self.router_proj,
-            router_bias: &self.router_bias,
-            experts_gate_up_bias: &self.gate_up_bias,
-            experts_down_bias: &self.down_bias,
-            router_scale: &[],
-            router_per_expert_scale: &[],
-            router_norm: &[],
-            router_norm_parameter_free: false,
-            router_input_scalar: 1.0,
-            pre_experts_norm: &self.pre_ffn_norm,
-            post_ffn1_norm: &[],
-            post_experts_norm: &[],
-            num_experts: self.experts,
-            top_k: self.top_k,
-            intermediate_size: self.inter,
-            gate_rule: self.gate_rule,
-        }
-    }
-}
-
-/// Stages this run omits, for marginal-cost profiling.
-///
-/// Every one of these is an operation the plan marks **optional**, so
-/// omitting it exercises a path the lowering already supports rather
-/// than a special diagnostic branch. The numbers are wrong by
-/// construction; the *time difference* is the measurement.
-///
-/// Ablation is used because this hardware supports counter sampling only
-/// at compute-pass boundaries (`AtDispatchBoundary` is false on M3), so
-/// per-dispatch GPU timestamps are unavailable. Splitting stages into
-/// separate encoders to get boundaries would change what can overlap;
-/// ablation leaves the schedule of everything that remains intact.
-#[derive(Clone, Copy, Default)]
-pub struct Ablation {
-    pub no_query_scale: bool,
-    pub no_rope: bool,
-    pub no_qk_norm: bool,
-    pub no_gate: bool,
-    pub no_post_norms: bool,
-}
-
-impl Ablation {
-    fn from_env() -> Self {
-        let on = |k: &str| std::env::var(k).is_ok();
-        Self {
-            no_query_scale: on("LARQL_ABLATE_QUERY_SCALE"),
-            no_rope: on("LARQL_ABLATE_ROPE"),
-            no_qk_norm: on("LARQL_ABLATE_QK_NORM"),
-            no_gate: on("LARQL_ABLATE_GATE"),
-            no_post_norms: on("LARQL_ABLATE_POST_NORMS"),
-        }
-    }
-
-    fn any(&self) -> bool {
-        self.no_query_scale || self.no_rope || self.no_qk_norm || self.no_gate || self.no_post_norms
-    }
-}
-
 /// A plan lowered onto the device, ready to step positions.
 pub struct LoweredSession<'a> {
     gpu: &'a MetalBackend,
@@ -239,140 +116,6 @@ pub struct LoweredSession<'a> {
     position: usize,
     /// Destination for resolved GPU timestamps, when profiling.
     ablate: Ablation,
-}
-
-/// Load one matrix operand as NVFP4 and hand it to the device.
-///
-/// The buffers are keyed on the `AlignedBytes` address, which lives for
-/// the session, so `lowering_weight` caches them and the weight is
-/// uploaded once rather than per position.
-fn resident_matrix(
-    gpu: &MetalBackend,
-    store: &OperandStore,
-    operand: &OperandRef,
-    format: WeightFormat,
-    keep: &mut Vec<LoadedWeight>,
-) -> Result<DeviceMatrix, VindexError> {
-    let rows = operand.shape.first().copied().unwrap_or(0);
-    let cols = operand.shape.get(1).copied().unwrap_or(0);
-    let loaded = load_weight(store, operand, format)?;
-    let m = match &loaded {
-        LoadedWeight::Nvfp4 {
-            packed,
-            scales,
-            tensor_scale,
-        } => DeviceMatrix {
-            packed: gpu.lowering_weight(packed.as_slice()),
-            scales: gpu.lowering_weight(scales.as_slice()),
-            tensor_scale: *tensor_scale,
-            format: WeightFormat::Nvfp4,
-            rows,
-            cols,
-        },
-        LoadedWeight::Mxfp4 { packed, scales } => DeviceMatrix {
-            packed: gpu.lowering_weight(packed.as_slice()),
-            scales: gpu.lowering_weight(scales.as_slice()),
-            tensor_scale: 1.0,
-            format: WeightFormat::Mxfp4,
-            rows,
-            cols,
-        },
-        LoadedWeight::F16(bytes) => DeviceMatrix {
-            packed: gpu.lowering_weight(bytes.as_slice()),
-            scales: gpu.lowering_weight(&[]),
-            tensor_scale: 1.0,
-            format: WeightFormat::F16,
-            rows,
-            cols,
-        },
-        _ => {
-            return Err(VindexError::Parse(format!(
-                "operand `{}`: unsupported lowering format {format:?}",
-                operand.tensor
-            )))
-        }
-    };
-    // The device buffers alias these allocations, so the session owns
-    // them for its lifetime.
-    keep.push(loaded);
-    Ok(m)
-}
-
-/// Upload an optional f32 vector operand (a bias or the sink logits) to
-/// the device, or `None` when the plan carries none.
-fn resident_vector(
-    gpu: &MetalBackend,
-    store: &OperandStore,
-    operand: Option<&OperandRef>,
-) -> Result<Option<DeviceBuffer>, VindexError> {
-    match operand {
-        Some(op) => {
-            let v = store.load(op)?;
-            let buf = gpu
-                .lowering_upload(&v)
-                .ok_or_else(|| VindexError::Parse("vector operand upload failed".into()))?;
-            Ok(Some(buf))
-        }
-        None => Ok(None),
-    }
-}
-
-/// The `inv_freq` map key for a rotary policy — distinct per (theta,
-/// scaled-or-plain) so YaRN and plain rope at the same base never share a
-/// table; `None` for NoPE.
-fn rope_table_key(position: &PositionPolicy) -> Option<u64> {
-    match position {
-        PositionPolicy::Rope { theta } => Some(theta.to_bits()),
-        // Fold the yarn block into the key so two different blocks (or a
-        // block vs plain rope) at one theta get their own tables. The
-        // block's f64 fields hash deterministically.
-        PositionPolicy::Yarn { theta, scaling } => {
-            let mut h = std::collections::hash_map::DefaultHasher::new();
-            use std::hash::{Hash, Hasher};
-            theta.to_bits().hash(&mut h);
-            scaling.factor.to_bits().hash(&mut h);
-            scaling.beta_fast.to_bits().hash(&mut h);
-            scaling.beta_slow.to_bits().hash(&mut h);
-            scaling
-                .original_max_position_embeddings
-                .to_bits()
-                .hash(&mut h);
-            scaling.truncate.hash(&mut h);
-            Some(h.finish() | 1)
-        }
-        PositionPolicy::None => None,
-    }
-}
-
-/// The inverse-frequency table for a rotary policy, matching the
-/// interpreter kernel exactly: plain `theta^(-2i/d)` for rope, the YaRN
-/// ramp for a scaled layer.
-fn rope_inv_freq_table(position: &PositionPolicy, head_dim: usize) -> Vec<f32> {
-    match position {
-        PositionPolicy::Rope { theta } => (0..head_dim / 2)
-            .map(|i| theta.powf(-2.0 * i as f64 / head_dim as f64) as f32)
-            .collect(),
-        PositionPolicy::Yarn { theta, scaling } => {
-            let (inv_freq, _amplitude) =
-                larql_vindex::format::vindex3::opplan::exec::kernels::yarn_frequencies(
-                    scaling, head_dim, *theta,
-                );
-            inv_freq.iter().map(|f| *f as f32).collect()
-        }
-        PositionPolicy::None => Vec::new(),
-    }
-}
-
-fn resident_norm(
-    gpu: &MetalBackend,
-    store: &OperandStore,
-    op: &NormOp,
-) -> Result<(DeviceBuffer, f32, f32), VindexError> {
-    let w = store.load(&op.weight)?;
-    let buf = gpu
-        .lowering_upload(&w)
-        .ok_or_else(|| VindexError::Parse("norm weight upload failed".into()))?;
-    Ok((buf, op.eps as f32, op.weight_offset))
 }
 
 impl<'a> LoweredSession<'a> {
@@ -885,205 +628,6 @@ impl<'a> LoweredSession<'a> {
     }
 }
 
-/// Build a layer's resident FFN: dense matrices, or the routed expert
-/// bank loaded into registered regions with its MoE machinery.
-fn build_ffn(
-    gpu: &MetalBackend,
-    store: &OperandStore,
-    layer: &LayerPlan,
-    formats: WeightFormats,
-    keep: &mut Vec<LoadedWeight>,
-) -> Result<FfnResident, VindexError> {
-    if let Some(op) = layer.ffn.routed() {
-        return Ok(FfnResident::Routed(Box::new(build_routed(
-            gpu, store, layer, op,
-        )?)));
-    }
-    let dense = dense_ffn(layer)?;
-    Ok(FfnResident::Dense {
-        gate: resident_matrix(
-            gpu,
-            store,
-            dense
-                .gate
-                .as_ref()
-                .ok_or_else(|| VindexError::Parse("lowering requires a gated FFN".into()))?,
-            formats.ffn,
-            keep,
-        )?,
-        up: resident_matrix(gpu, store, &dense.up, formats.ffn, keep)?,
-        down: resident_matrix(gpu, store, &dense.down, formats.ffn, keep)?,
-    })
-}
-
-/// Load a routed layer's expert bank into page-aligned, region-registered
-/// buffers and build its MoE scratch + descriptor table. The expert bytes
-/// are bound zero-copy through the same registered-region path the served
-/// `--routed-from` run uses — never copied per token.
-fn build_routed(
-    gpu: &MetalBackend,
-    store: &OperandStore,
-    layer: &LayerPlan,
-    op: &larql_vindex::format::vindex3::opplan::RoutedFfnOp,
-) -> Result<RoutedLayer, VindexError> {
-    use larql_vindex::format::vindex3::opplan::exec::weights::AlignedBytes;
-    // Only the packed-MXFP4 gpt-oss bank lowers today (native experts,
-    // split scales); any other routed shape refuses rather than guess.
-    if op.expert_format != larql_models::ExpertFormat::PackedMxfp4 {
-        return Err(VindexError::Parse(format!(
-            "layer {}: the Metal lowering serves only PackedMxfp4 routed experts, not {:?}",
-            layer.layer, op.expert_format
-        )));
-    }
-    let hidden = op.router.shape.get(1).copied().unwrap_or(0);
-    let experts = op.experts;
-    let inter = op.expert_intermediate_size;
-
-    // Packed blocks and scales into aligned, registered banks.
-    let aligned = |operand: &larql_vindex::format::vindex3::opplan::OperandRef| -> Result<AlignedBytes, VindexError> {
-        let raw = store.load_raw(operand)?;
-        Ok(AlignedBytes::from_bytes(&raw.bytes))
-    };
-    let gate_up_blocks = aligned(&op.gate_up.weights)?;
-    let gate_up_scales = aligned(op.gate_up.scales.as_ref().ok_or_else(|| {
-        VindexError::Parse(format!(
-            "layer {}: routed gate_up carries no scales",
-            layer.layer
-        ))
-    })?)?;
-    let down_blocks = aligned(&op.down.weights)?;
-    let down_scales = aligned(op.down.scales.as_ref().ok_or_else(|| {
-        VindexError::Parse(format!(
-            "layer {}: routed down carries no scales",
-            layer.layer
-        ))
-    })?)?;
-    for (bank, what) in [
-        (&gate_up_blocks, "gate_up blocks"),
-        (&gate_up_scales, "gate_up scales"),
-        (&down_blocks, "down blocks"),
-        (&down_scales, "down scales"),
-    ] {
-        if !gpu.lowering_register_region(bank.as_slice()) {
-            return Err(VindexError::Parse(format!(
-                "layer {}: could not register the routed {what} region (not page-aligned)",
-                layer.layer
-            )));
-        }
-    }
-    let gu_expert_bytes = gate_up_blocks.logical_len() / experts;
-    let gu_scale_bytes = gate_up_scales.logical_len() / experts;
-    let dn_expert_bytes = down_blocks.logical_len() / experts;
-    let dn_scale_bytes = down_scales.logical_len() / experts;
-
-    let f32_or_empty =
-        |o: Option<&larql_vindex::format::vindex3::opplan::OperandRef>| -> Result<Vec<f32>, VindexError> {
-            match o {
-                Some(op) => store.load(op),
-                None => Ok(Vec::new()),
-            }
-        };
-    let router_proj = store.load(&op.router)?;
-    let router_bias = f32_or_empty(op.router_bias.as_ref())?;
-    let gate_up_bias = f32_or_empty(op.gate_up.bias.as_ref())?;
-    let down_bias = f32_or_empty(op.down.bias.as_ref())?;
-    let pre_ffn_norm = store.load(&layer.pre_ffn_norm.weight)?;
-    let gate_rule = larql_compute::MoeGateRule::from_arch(op.gate_policy, op.activation);
-
-    let scratch = larql_compute_metal::MoeScratch::new_public_with_format(
-        gpu,
-        op.top_k,
-        hidden,
-        inter,
-        larql_compute::QuantFormat::MXFP4,
-        hidden,
-    );
-    // Build the descriptor table from a temporary `MoeLayerWeights`
-    // borrowing the freshly-loaded storage, before that storage moves
-    // into `RoutedLayer` — the table keeps only region buffers, no borrow.
-    let table = {
-        use larql_compute::{
-            MoeExpertScales, MoeFusedRowLayout, MoeRoutingPolicy, MoeWeightLayout, QuantFormat,
-        };
-        let moe = larql_compute::MoeLayerWeights {
-            experts_gate_up: expert_slices(&gate_up_blocks, gu_expert_bytes, experts),
-            experts_down: expert_slices(&down_blocks, dn_expert_bytes, experts),
-            routing_policy: MoeRoutingPolicy::top_k_then_softmax(),
-            weight_layout: MoeWeightLayout::unpadded(),
-            expert_scales: MoeExpertScales::Paired {
-                gate_up: expert_slices(&gate_up_scales, gu_scale_bytes, experts),
-                down: expert_slices(&down_scales, dn_scale_bytes, experts),
-            },
-            fused_row_layout: MoeFusedRowLayout::Interleaved,
-            expert_data_format: QuantFormat::MXFP4,
-            router_proj: &router_proj,
-            router_bias: &router_bias,
-            experts_gate_up_bias: &gate_up_bias,
-            experts_down_bias: &down_bias,
-            router_scale: &[],
-            router_per_expert_scale: &[],
-            router_norm: &[],
-            router_norm_parameter_free: false,
-            router_input_scalar: 1.0,
-            pre_experts_norm: &pre_ffn_norm,
-            post_ffn1_norm: &[],
-            post_experts_norm: &[],
-            num_experts: experts,
-            top_k: op.top_k,
-            intermediate_size: inter,
-            gate_rule,
-        };
-        if !gpu.lowering_moe_supported(&moe, &scratch) {
-            return Err(VindexError::Parse(format!(
-                "layer {}: the descriptor MoE path does not support this routed layer \
-                 (format/policy/geometry) — refusing before encode",
-                layer.layer
-            )));
-        }
-        gpu.lowering_moe_descriptor(layer.layer, &moe, inter, hidden)
-            .ok_or_else(|| {
-                VindexError::Parse(format!(
-                    "layer {}: expert operands did not resolve inside their registered regions",
-                    layer.layer
-                ))
-            })?
-    };
-    Ok(RoutedLayer {
-        gate_up_blocks,
-        gate_up_scales,
-        down_blocks,
-        down_scales,
-        router_proj,
-        router_bias,
-        gate_up_bias,
-        down_bias,
-        pre_ffn_norm,
-        gu_expert_bytes,
-        gu_scale_bytes,
-        dn_expert_bytes,
-        dn_scale_bytes,
-        experts,
-        top_k: op.top_k,
-        inter,
-        gate_rule,
-        table,
-        scratch,
-        eps: layer.pre_ffn_norm.eps as f32,
-    })
-}
-
-/// The dense FFN op of a layer the lowering has already admitted (routed
-/// layers are refused in `new`, so this only fails on a plan that changed
-/// under us).
-fn dense_ffn(layer: &LayerPlan) -> Result<&FfnOp, VindexError> {
-    layer.ffn.dense().ok_or_else(|| {
-        VindexError::Parse(format!(
-            "layer {} carries a routed FFN the lowering does not execute (A-9.4)",
-            layer.layer
-        ))
-    })
-}
-
 fn argmax_of(logits: &[f32]) -> u32 {
     logits
         .iter()
@@ -1099,67 +643,6 @@ fn argmax_of(logits: &[f32]) -> u32 {
             },
         )
         .0 as u32
-}
-
-/// Teacher-force `tokens` through the lowering, capturing every layer's
-/// output per position, and write the planes + manifest a `shannon
-/// layer-diff` consumes — byte-compatible with `vindex3 exec --dump-layers`
-/// on the interpreter backends, so the two can be diffed directly.
-fn dump_lowered(
-    session: &mut LoweredSession<'_>,
-    tokens: &[u32],
-    plan: &ComponentOpPlan,
-    container: &std::path::Path,
-    label: &str,
-    dir: &std::path::Path,
-) -> Result<(), Box<dyn std::error::Error>> {
-    use super::super::shannon_trace::dump::{
-        plane_name, LayerDumpManifest, MANIFEST_NAME, PLANE_DTYPE,
-    };
-    let hidden = session.hidden;
-    let num_layers = plan.layers.len();
-    let seq = tokens.len();
-    std::fs::create_dir_all(dir)?;
-    // Row-major [seq, hidden] planes: plane 0 = embeddings, plane i+1 =
-    // layer i's post-FFN-residual hidden.
-    let mut planes: Vec<Vec<f32>> = vec![Vec::with_capacity(seq * hidden); num_layers + 1];
-    for (pos, &token) in tokens.iter().enumerate() {
-        let (_logits, embedding, layers_out) = session.step_capturing(token)?;
-        planes[0].extend_from_slice(&embedding);
-        for (i, row) in layers_out.into_iter().enumerate() {
-            planes[i + 1].extend_from_slice(&row);
-        }
-        eprintln!("captured position {}/{seq}", pos + 1);
-    }
-    let plane_names: Vec<String> = (0..=num_layers).map(plane_name).collect();
-    for (name, data) in plane_names.iter().zip(&planes) {
-        use std::io::Write;
-        let mut f = std::io::BufWriter::new(std::fs::File::create(dir.join(name))?);
-        for v in data {
-            f.write_all(&v.to_le_bytes())?;
-        }
-        f.flush()?;
-    }
-    let manifest = LayerDumpManifest {
-        engine: format!("vindex3-metal-lowered-{label}"),
-        model: container.display().to_string(),
-        num_layers,
-        seq_len: seq,
-        hidden_size: hidden,
-        token_ids: tokens.to_vec(),
-        planes: plane_names,
-        dtype: PLANE_DTYPE.to_string(),
-    };
-    std::fs::write(
-        dir.join(MANIFEST_NAME),
-        serde_json::to_string_pretty(&manifest)?,
-    )?;
-    eprintln!(
-        "wrote {} planes + manifest to {}",
-        num_layers + 1,
-        dir.display()
-    );
-    Ok(())
 }
 
 /// Run the plan through the lowering and report the final position's
